MAINT Extract scenario progress read model
